# public/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas App API
  description: |
    API REST de la aplicación de finanzas personales (Next.js App Router + Prisma + PostgreSQL).

    ## Autenticación
    La mayoría de los endpoints requieren un **access token JWT** en el header
    `Authorization: Bearer <token>`. El token se obtiene en `/auth/login` o `/auth/register`.
    El **refresh token** se entrega como cookie `HttpOnly` (`refresh_token`) y se usa en `/auth/refresh`.

    ## Convenciones
    - Montos monetarios: número decimal (la API los serializa a veces como string desde PostgreSQL).
    - Fechas: ISO 8601 (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - Errores de validación → `422` con `{ error, details }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /split:
    post:
      tags: [Split]
      summary: Crear cobro compartido y link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [concept, totalAmount, bankDetails, participants]
              properties:
                concept:     { type: string, maxLength: 200 }
                totalAmount: { type: number }
                notes:       { type: string }
                bankDetails:
                  type: object
                  required: [bank]
                  properties:
                    bank:  { type: string }
                    clabe: { type: string }
                    alias: { type: string }
                    phone: { type: string }
                participants:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items:
                    type: object
                    required: [amount]
                    properties:
                      userId:     { type: string }
                      guestName:  { type: string }
                      guestEmail: { type: string, format: email }
                      amount:     { type: number }
      responses:
        '201': { description: Cobro creado (incluye shareUrl) }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /split/{token}/pay:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string } }
    post:
      tags: [Split]
      summary: Marcar pago de un participante (auth o invitado)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId: { type: string }
                guestName:     { type: string }
                guestEmail:    { type: string, format: email }
      responses:
        '200': { description: Pago registrado }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ─────────────────────── NOTIFICATIONS ──────────────────
...
  /sync:
    post:
      tags: [Sync]
      summary: Sincronización offline por lotes
      description: Procesa hasta 50 acciones encoladas por el Service Worker (Background Sync).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actions]
              properties:
                actions:
                  type: array
                  maxItems: 50
                  items:
                    type: object
                    required: [clientId, method, resource]
                    properties:
                      clientId: { type: string }
                      method:   { type: string, enum: [POST, PUT, PATCH, DELETE] }
                      resource: { type: string, enum: [transactions, accounts] }
                      payload:  {}
      responses:
        '200': { description: Resultado por acción }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token JWT obtenido en /auth/login o /auth/register'
  responses:
...
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string, example: Datos inválidos }
              details: { type: object, additionalProperties: true }
...